Renames find and check event function.
