+++ lib/assertions/2wp.js
@@ -137,7 +137,7 @@ const assertRejectedPeginEvent = (rejectedPeginTx, expectedRejectionReason, expe
   expect(outpointValues.every(value => value in federationUtxoValues)).to.be.true;
 }
 
-const findAndCheckPeginBtcEventAtBlock = async (rskTxHelper, atBlock, expectedEvent) => {
+const findAndCheckPeginBtcEventInBlock = async (rskTxHelper, atBlock, expectedEvent) => {
   const peginBtcEvent = await rskUtils.findEventInBlock(rskTxHelper, expectedEvent.name, atBlock);
   expect(peginBtcEvent).to.exist;
   peginBtcEvent.arguments.receiver = peginBtcEvent.arguments.receiver.toLowerCase();
@@ -151,5 +151,5 @@ module.exports = {
   }),
   assertCallToPegoutBatchingBridgeMethods,
   assertRejectedPeginEvent,
-  findAndCheckPeginBtcEventAtBlock,
+  findAndCheckPeginBtcEventInBlock,
 };
