+++ tests/01_01_01-2wp.js
@@ -1,3 +1,3 @@
-const twoWpTests = require('../lib/tests/2wp-new');
+const twoWpTests = require('../lib/tests/2wp');
 
 twoWpTests.execute('BTC <=> RSK 2WP', () => Runners.hosts.federate.host);
